Move clash() to npc_combat; write test
